+++ core/prelude-global-keybindings.el
@@ -101,8 +101,17 @@
 
 (global-set-key (kbd "C-=") 'er/expand-region)
 
-(global-set-key (kbd "C-c v") 'avy-goto-word-or-subword-1)
+;; recommended avy keybindings
+(global-set-key (kbd "C-:") 'avy-goto-char)
+(global-set-key (kbd "C-'") 'avy-goto-char-2)
+(global-set-key (kbd "M-g f") 'avy-goto-line)
+(global-set-key (kbd "M-g w") 'avy-goto-word-1)
+(global-set-key (kbd "M-g e") 'avy-goto-word-0)
+
+;; additional avy keybindings
+(global-set-key (kbd "s-,") 'avy-goto-char)
 (global-set-key (kbd "s-.") 'avy-goto-word-or-subword-1)
+(global-set-key (kbd "C-c v") 'avy-goto-word-or-subword-1)
 
 ;; improved window navigation with ace-window
 (global-set-key (kbd "s-w") 'ace-window)